[ARIES-1252] Initial proposal to support custom content in .esa files
Custom content in .esa files is handled via Whiteboard service handlers.
To add a handler for custom content, register a
org.apache.aries.subsystem.ContentHandler service. This service will be
called back for install/start/stop/actions to be performed. (This initial
implementation only actually handles install right now.)
The service must be registered with the service property
  org.aries.subsystem.contenthandler.type
to declare the content type that it should handle. This allows to
register handlers for content types different than the default
osgi.bundle/osgi.fragment and osgi.subsystem.* ones.

For more details take a look at the CustomContentHandlerTest itest. This
creates an .esa file that contains a bundle and also a custom content
file called 'custom1.sausages'. The custom content is declared in the
Subsystem-Content header via the standard 'type' attribute. To link it
with a resource in the .esa also provide the embedded-resource
attribute, e.g.:
Subsystem-Content:
  customContent1;embedded-resource=custom1.sausages;type=foo.sausages

Using a proper resource type (rather than directly using the file name
extension) makes it possible to also provide this custom resource from a
repository if needed.

The CustomContentHandlerTest has a test content handler that is called
back with the content of the resource, its name, the current subsystem
and the coordination.
Providing the coordination allows the custom handler to define a
compensation in case the installation fails elsewhere, or fail the
installation itself.

Note that this implementation is not finished and also needs more tests.
Currently it's just a proposal...